Fixes warnings

+++ CircleBulletPattern.cpp
@@ -6,7 +6,7 @@ void CircleBulletPattern::createBullet(int currentFrame, bool rightTurretActive,
     Vector2 rightTurretPosition, Vector2 leftTurretPosition, std::vector<GameStatus::BulletInfo> *bullets) {
     const int numBulletSources = 10;
     for (int i = 0; i <= numBulletSources; ++i) {
-        float angle = (i * M_PI / numBulletSources);
+        float angle = (float)(i * M_PI / numBulletSources);
         if (rightTurretActive) {
             bullets->push_back(GameStatus::BulletInfo({ cos(angle) * this->bulletSpeed, sin(angle) * this->bulletSpeed }, rightTurretPosition));
         }
